DebugLib: Allow -Wunneeded-internal-declaration with clang
Provides a variant of the DEBUG macro for clang when MDEPKG_NDEBUG is
defined, which uses but discards the contained expression. This means
clang can tell that it has optimised away variable usage as part of
valid debug patterns (such as a STATIC variable which is only used in
DEBUG statements) - rather than just seeing such variables as
completely unused - therefore we can keep
-Wunneeded-internal-declaration (as part of -Wall) to warn about
mistakenly genuinely unused variables elsewhere.

Note that the _Pragma in the clang/gcc variant is to temporarily
suppress the warning about `(VOID) Expression;`, whilst the presence
of `(VOID) Expression;` (once allowed) is what prevents the unwanted
warning about unneeded variables.

Signed-off-by: Mikhail Krichanov <mikhailkrichanov@gmail.com>
Signed-off-by: Mike Beaton <mjsbeaton@gmail.com>