MdePkg: Fix incorrect signature in PEI and DXE COPY_MEM and SET_MEM typedefs
In all four cases fields of the given types are initialised from methods
matching the CopyMem and SetMem declarations in BaseMemoryLib.h, which
have return values.

Fixing this is required to be able to compile with cast-function-type-mismatch
warning enabled for Xcode, but appears to be a genuine unwanted mismatch in
any case.

Signed-off-by: Mike Beaton <mjsbeaton@gmail.com>