TEMP - CryptoPkg: Fix compilation on XCODE5 toolchain
It is not so obvious that this works, because in addition
to the openssl library, __APPLE__ is also checked in two
other locations under CryptoPkg/Library/OpensslLib:

./OpensslGen/AARCH64-GCC/crypto/arm64cpuid.S:#ifdef  __APPLE__
./OpensslGen/include/openssl/bio.h:    && !defined(__APPLE__)

Plus also multiple locations in openssl, and it is not obvious
that we actually want to disable the operation of all of these
either, if you look at what they do.

This is unlike the similar fix in MdePkg/Library/BaseFdtLib
where __APPLE__ fortunately occurs only in the library
libfdt which we want to affect.

If it does work:

---

Required after EDK 2 update to openssl 3.5.1.

Introduced in openssl by:
https://github.com/openssl/openssl/commit/92c242e8ac26e1d4cb692c5258d0aefa14e5de84

Signed-off-by: Mike Beaton <mjsbeaton@gmail.com>