.azurepipelines: Upgrade NodeTool@0 to UseNode@1
In addition to upgrading from CmdLine@1 to CmdLine@2 in
the previous commit, there is one further Azure Pipelines
task in use in CI which is as of now has a newer version
available:

https://learn.microsoft.com/en-us/azure/devops/pipelines/tasks/reference/node-tool-v0
https://learn.microsoft.com/en-us/azure/devops/pipelines/tasks/reference/use-node-v1

so we upgrade that in this commit.

Note that `checkLatest` is described in exactly the same terms
in both versions, so it is assumed that the commented out line
`#checkLatest: false # Optional` is valid to leave in place for
commenting in if required.

Signed-off-by: Mike Beaton <mjsbeaton@gmail.com>