.azurepipelines: Run non-package-build CI only in GCC and VS2022
We remove NO-TARGET and NOOPT builds from calls to stuart_ci_build
except in GCC and VS2022.

This amounts to removing:
 - Unit tests (NOOPT)
 - Uncrustify, spell check, etc. (NO-TARGET)

Note that in PR gate CI, NOOPT is never built in as a normal
package build (see .pytool/Plugin/CompilerPlugin, which only runs on
DEBUG and RELEASE). Instead NOOPT is used for building unit tests
(see .pytool/Plugin/HostUnitTestCompilerPlugin).

Signed-off-by: Mike Beaton <mjsbeaton@gmail.com>